+++ lib/pages/eventsPage.dart
@@ -4,7 +4,10 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 class EventsPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return new Scaffold(
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Events'),
+      ),
       body: Center(
         child: Padding(
           padding: const EdgeInsets.only(bottom: 100.0),
